+++ tests/conftest.py
@@ -1,4 +1,5 @@
 import os
+import subprocess
 from logging import INFO
 
 import pytest
@@ -58,3 +59,19 @@ def prepare_authorization_dict():
             'org_sp_fqdn': 'org.sp3.co.jp'
         }
     }
+
+
+@pytest.fixture
+def switch_to_sentinel():
+    """Switch to sentinel for testing
+    
+    This fixture stops the redis container and starts the sentinel container
+    for testing purposes. It will switch back to the redis container after the
+    test is complete.
+    """
+    subprocess.run(['docker', 'compose', 'stop', 'redis'])
+    subprocess.run(['docker', 'compose', '-f', '../../docker-compose-sentinel.yml', 'up', '-d'])
+    yield
+    subprocess.run(['docker', 'compose', '-f', '../../docker-compose-sentinel.yml', 'down'])
+    subprocess.run(['docker', 'compose', 'start', 'redis'])
+
