+++ weko_group_cache_db/tasks.py
@@ -1,12 +1,11 @@
-from celery import shared_task
-from celery.utils.log import get_task_logger
+import logging
 
-from config import config, messages
+from . import config
+from . import messages
 from .utils import set_groups
 
-logger = get_task_logger(__name__)
+logger = logging.getLogger(__name__)
 
-@shared_task
 def create_set_groups_task():
     """Create set_groups_task for all sp in config.
     """
@@ -18,13 +17,12 @@ def create_set_groups_task():
             if not config.SP_AUTHORIZATION_DICT[org].get('org_sp_fqdn'):
                 logger.error(messages.FQDN_NOT_DEFINED.format(org))
                 continue
-            set_groups_task.delay(config.SP_AUTHORIZATION_DICT[org]['org_sp_fqdn'])
+            set_groups_task(config.SP_AUTHORIZATION_DICT[org]['org_sp_fqdn'])
     except Exception as ex:
         logger.error(messages.CREATE_TASK_FAILED, exc_info=True)
         raise ex
     logger.info(messages.CREATE_TASK_SUCCESS)
 
-@shared_task
 def set_groups_task(fqdn):
     """Get groups from Gakunin API and set to Redis
